change way of parsing data
instead of configuring struct layout first and then parsing according
to that layout, just parse as code on the fly. This makes dynamic
stuff much easier (like fields changing type depending on other
field values, or dynamically sized structs) and it's just less of
a mess. This is makes it possible to give names to structures based
on its fields, so also do that.